+++ ftc/src/main/java/org/psilynx/psikit/ftc/wrappers/VoltageSensorWrapper.kt
@@ -2,6 +2,7 @@ package org.psilynx.psikit.ftc.wrappers
 
 import com.qualcomm.robotcore.hardware.HardwareDevice
 import com.qualcomm.robotcore.hardware.VoltageSensor
+import org.psilynx.psikit.ftc.FtcLogTuning
 import org.psilynx.psikit.core.LogTable
 
 class VoltageSensorWrapper(
@@ -9,6 +10,7 @@ class VoltageSensorWrapper(
 ) : VoltageSensor, HardwareInput<VoltageSensor> {
 
     private var _voltage = 0.0
+    private var _sampledThisLoop = false
     private var _deviceName = "MockVoltageSensor"
     private var _version = 1
     private var _connectionInfo = ""
@@ -17,19 +19,38 @@ class VoltageSensorWrapper(
     override fun new(wrapped: VoltageSensor?) = VoltageSensorWrapper(wrapped)
 
     override fun toLog(table: LogTable) {
-        device!!
+        val d = device
 
-        _voltage = device.voltage
-        _deviceName = device.deviceName
-        _version = device.version
-        _connectionInfo = device.connectionInfo
-        _manufacturer = device.manufacturer
+        if (d != null) {
+            _deviceName = d.deviceName
+            _version = d.version
+            _connectionInfo = d.connectionInfo
+            _manufacturer = d.manufacturer
+        }
 
-        table.put("voltage", voltage)
+        if (FtcLogTuning.bulkOnlyLogging) {
+            if (_sampledThisLoop) {
+                table.put("voltage", _voltage)
+            }
+            table.put("voltage/sampled", _sampledThisLoop)
+            _sampledThisLoop = false
+            return
+        }
+
+        if (d != null) {
+            _voltage = d.voltage
+            _sampledThisLoop = true
+        }
+
+        if (_sampledThisLoop) {
+            table.put("voltage", _voltage)
+        }
+        table.put("voltage/sampled", _sampledThisLoop)
         table.put("deviceName", deviceName)
         table.put("version", version)
         table.put("connectionInfo", connectionInfo)
         table.put("manufacturer", manufacturer)
+        _sampledThisLoop = false
 
 
     }
@@ -40,9 +61,17 @@ class VoltageSensorWrapper(
         _version = table.get("version", 1)
         _connectionInfo = table.get("connectionInfo", "")
         _manufacturer = table.get("manufacturer", HardwareDevice.Manufacturer.Other)
+        _sampledThisLoop = table.get("voltage/sampled", false)
     }
 
-    override fun getVoltage() = _voltage
+    override fun getVoltage(): Double {
+        val d = device
+        if (d != null) {
+            _voltage = d.voltage
+            _sampledThisLoop = true
+        }
+        return _voltage
+    }
 
     override fun getDeviceName() = _deviceName
     override fun getVersion() = _version